CodeSnippets content/style/animation/placeholders. Repositories and ViewModels refactoring.
1. CodeSnippets content now resides in local DB as JS code, ready to be
inserted to webView. This was made to be able to use CodeSnippets
without internet connection. The same was made for CodeSnippets CSS style.
A new entity and new database were created to store CSS style. Initial
animation was made to show CodeSnippets after they loaded. Placeholders
were added to show "loading" and "no internet connection" status of
CodeSnippets.

2. All data handling logic was moved from ViewModel to Repositories,
as it should be. Some code was added and some code was deleted, accordingly.

! All changes in this commit should be refactored to better code quality.